requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
pydantic-settings==2.6.1
requests==2.32.3
groq==0.11.0
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1